Restructured the app structure to look closely to the original sample for the deployment test
