--- src/__init__.py
@@ -1,3 +1,0 @@
-from crew_base import KeboolaInsightsCrew, get_status
-
-__all__ = ['KeboolaInsightsCrew', 'get_status']
